added selfhosted-operator mode for test

+++ test/e2e/api_server_test.go
@@ -28,19 +28,25 @@ var _ = Describe("API server", func() {
 	})
 
 	JustBeforeEach(func() {
-		By("Starting Kubed")
-		stopCh = make(chan struct{})
-		err := f.RunKubed(stopCh, clusterConfig)
-		Expect(err).NotTo(HaveOccurred())
+		if f.SelfHostedOperator {
+			f.RestartKubedOperator(&clusterConfig)
+		} else {
+			By("Starting Kubed")
+			stopCh = make(chan struct{})
+			err := f.RunKubed(stopCh, clusterConfig)
+			Expect(err).NotTo(HaveOccurred())
 
-		By("Waiting for API server to be ready")
-		root.EventuallyAPIServerReady().Should(Succeed())
-		time.Sleep(time.Second * 5)
+			By("Waiting for API server to be ready")
+			root.EventuallyAPIServerReady().Should(Succeed())
+			time.Sleep(time.Second * 5)
+		}
 	})
 
 	AfterEach(func() {
-		close(stopCh)
-		os.RemoveAll(filepath.Join("/tmp", "indices"))
+		if !f.SelfHostedOperator {
+			close(stopCh)
+			os.RemoveAll(filepath.Join("/tmp", "indices"))
+		}
 	})
 
 	Describe("Search object", func() {
